requirements.txt
# PetClassifier - Dog and Cat Classifier
# Project dependencies

# PyTorch and torchvision for deep learning
torch>=1.10.0
torchvision>=0.11.0

# Image processing
Pillow>=9.0.0

# Data analysis and metrics (for training notebooks)
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0

# Optional: For development and notebooks
jupyter>=1.0.0
notebook>=6.4.0
